fix: condition for have no licence

+++ hooks/post_gen_project.py
@@ -1,5 +1,6 @@
 import os
 import shutil
 
-open('LICENSE.md', 'w').write(open('licenses/{{ cookiecutter.license }}','r').read())
-shutil.rmtree('licenses')
+if '{{ cookiecutter.license }}' != 'Unlicensed':
+    open('LICENSE.md', 'w').write(open('licenses/{{ cookiecutter.license }}','r').read())
+    shutil.rmtree('licenses')
